third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEOffsetElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dropShadowElement.in1.baseVal is "SourceAlpha"
PASS dropShadowElement.in1.baseVal is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
